fourth task

+++ src/main/java/lang/print/gaps/task4/SettingTypes.java
@@ -7,7 +7,7 @@ public class SettingTypes {
         boolean third = 5 > 9;
         double forth = 12.90434323324342;
         float fifth = 12.3F;
-        short sixth = 1;
-        int seventh = 129;
+        byte sixth = 1;
+        short seventh = 129;
     }
 }
